% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byContinent.R
\name{byContinent}
\alias{byContinent}
\title{Output a vector containing the continent of the country}
\usage{
byContinent(x, y)
}
\arguments{
\item{x}{A vector of country names and percentages.}

\item{y}{A vector of percentages.}
}
\value{
A data frame with two columns: continent and average percentage.
}
\description{
Output a vector containing the continent of the country
}
\examples{
data(cousin)
byContinent(cousin$Country, cousin$Percent)
}

