% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cousin}
\alias{cousin}
\title{Cousin Marriage Data}
\format{A data frame with 70 rows and 2 variables:
\describe{
  \item{Country}{country}
  \item{Percent}{percentage of marriages that are consanguineous}
  ...
}}
\source{
\url{https://github.com/fivethirtyeight/data/tree/master/cousin-marriage}
}
\usage{
cousin
}
\description{
A dataset containing the percentage of marriages that are consanguineous for various countries
}
\keyword{datasets}

